LPS-153267 Moving base prop types to main Actions component
